New Y2Storage namespace
