+++ src/lib/y2storage.rb
@@ -1,5 +1,3 @@
-#!/usr/bin/env ruby
-#
 # encoding: utf-8
 
 # Copyright (c) [2016] SUSE LLC
@@ -21,31 +19,16 @@
 # To contact SUSE LLC about this file by physical or electronic mail, you may
 # find current contact information at www.suse.com.
 
-require "storage"
+require "y2storage/boot_requirements_checker"
+require "y2storage/devices_lists"
+require "y2storage/disk_analyzer"
 require "y2storage/disk_size"
-
-module Y2Storage
-  module Refinements
-    # Refinements for Partition adding some virtual attributes, mainly used
-    # to make the rspec tests more readable
-    module PartitionAttributes
-      refine ::Storage::Partition do
-        # First mounpoint
-        def mountpoint
-          filesystem.mountpoints.first
-        end
-
-        # Label of the filesystem
-        def label
-          filesystem.label
-        end
-
-        # UUID of the filesystem
-        def uuid
-          filesystem.uuid
-        end
-
-      end
-    end
-  end
-end
+require "y2storage/fake_device_factory"
+require "y2storage/free_disk_space"
+require "y2storage/planned_volume"
+require "y2storage/planned_volumes_list"
+require "y2storage/proposal"
+require "y2storage/proposal_settings"
+require "y2storage/refinements"
+require "y2storage/storage_manager"
+require "y2storage/yaml_writer"
