+++ src/lib/y2storage/boot_requirements_checker.rb
@@ -22,60 +22,58 @@
 # find current contact information at www.suse.com.
 
 require "yast"
-require "storage/boot_requirements_strategies"
-require "storage/storage_manager"
+require "y2storage/boot_requirements_strategies"
+require "y2storage/storage_manager"
 
-module Yast
-  module Storage
-    #
-    # Class that can check requirements for the different kinds of boot
-    # partition: /boot, EFI-boot, PReP.
-    #
-    # TO DO: Check with arch maintainers if the requirements are correct.
-    #
-    # See also
-    # https://github.com/yast/yast-bootloader/blob/master/SUPPORTED_SCENARIOS.md
-    #
-    class BootRequirementsChecker
-      include Yast::Logger
+module Y2Storage
+  #
+  # Class that can check requirements for the different kinds of boot
+  # partition: /boot, EFI-boot, PReP.
+  #
+  # TO DO: Check with arch maintainers if the requirements are correct.
+  #
+  # See also
+  # https://github.com/yast/yast-bootloader/blob/master/SUPPORTED_SCENARIOS.md
+  #
+  class BootRequirementsChecker
+    include Yast::Logger
 
-      class Error < RuntimeError
+    class Error < RuntimeError
+    end
+
+    def initialize(settings, disk_analyzer)
+      @settings = settings
+      @disk_analyzer = disk_analyzer
+    end
+
+    def needed_partitions
+      strategy.needed_partitions
+    rescue BootRequirementsStrategies::Error => error
+      raise Error, error.message
+    end
+
+  protected
+
+    attr_reader :settings
+    attr_reader :disk_analyzer
+
+    def arch
+      @arch ||= StorageManager.instance.arch
+    end
+
+    def strategy
+      return @strategy unless @strategy.nil?
+
+      if arch.x86? && arch.efiboot?
+        @strategy = BootRequirementsStrategies::UEFI.new(settings, disk_analyzer)
+      elsif arch.s390?
+        @strategy = BootRequirementsStrategies::ZIPL.new(settings, disk_analyzer)
+      elsif arch.ppc?
+        @strategy = BootRequirementsStrategies::PReP.new(settings, disk_analyzer)
       end
 
-      def initialize(settings, disk_analyzer)
-        @settings = settings
-        @disk_analyzer = disk_analyzer
-      end
-
-      def needed_partitions
-        strategy.needed_partitions
-      rescue BootRequirementsStrategies::Error => error
-        raise Error, error.message
-      end
-
-    protected
-
-      attr_reader :settings
-      attr_reader :disk_analyzer
-
-      def arch
-        @arch ||= StorageManager.instance.arch
-      end
-
-      def strategy
-        return @strategy unless @strategy.nil?
-
-        if arch.x86? && arch.efiboot?
-          @strategy = BootRequirementsStrategies::UEFI.new(settings, disk_analyzer)
-        elsif arch.s390?
-          @strategy = BootRequirementsStrategies::ZIPL.new(settings, disk_analyzer)
-        elsif arch.ppc?
-          @strategy = BootRequirementsStrategies::PReP.new(settings, disk_analyzer)
-        end
-
-        # Fallback to Legacy as default
-        @strategy ||= BootRequirementsStrategies::Legacy.new(settings, disk_analyzer)
-      end
+      # Fallback to Legacy as default
+      @strategy ||= BootRequirementsStrategies::Legacy.new(settings, disk_analyzer)
     end
   end
 end
