+++ src/lib/y2storage/boot_requirements_strategies/prep.rb
@@ -21,59 +21,53 @@
 # To contact SUSE LLC about this file by physical or electronic mail, you may
 # find current contact information at www.suse.com.
 
-require "yast"
-require "storage/boot_requirements_strategies/base"
-require "storage/planned_volume"
-require "storage/planned_volumes_list"
-require "storage/disk_size"
+require "y2storage/boot_requirements_strategies/base"
 
-module Yast
-  module Storage
-    module BootRequirementsStrategies
-      # Strategy to calculate boot requirements in systems using PReP
-      class PReP < Base
-        def needed_partitions
-          volumes = super
-          volumes << prep_volume if prep_partition_needed? && prep_partition_missing?
-          volumes
-        end
+module Y2Storage
+  module BootRequirementsStrategies
+    # Strategy to calculate boot requirements in systems using PReP
+    class PReP < Base
+      def needed_partitions
+        volumes = super
+        volumes << prep_volume if prep_partition_needed? && prep_partition_missing?
+        volumes
+      end
 
-      protected
+    protected
 
-        def boot_partition_needed?
-          settings.use_lvm # TODO: || settings.encrypted
-        end
+      def boot_partition_needed?
+        settings.use_lvm # TODO: || settings.encrypted
+      end
 
-        def prep_partition_needed?
-          # no need of PReP partition in OPAL/PowerNV/Bare metal
-          !arch.ppc_power_nv?
-        end
+      def prep_partition_needed?
+        # no need of PReP partition in OPAL/PowerNV/Bare metal
+        !arch.ppc_power_nv?
+      end
 
-        def prep_partition_missing?
-          partitions = disk_analyzer.prep_partitions[settings.root_device]
-          partitions.nil? || partitions.empty?
-        end
+      def prep_partition_missing?
+        partitions = disk_analyzer.prep_partitions[settings.root_device]
+        partitions.nil? || partitions.empty?
+      end
 
-        def prep_volume
-          vol = PlannedVolume.new(nil)
-          # So far we are always using msdos partition ids
-          vol.partition_id = ::Storage::ID_PPC_PREP
-          vol.min_disk_size = DiskSize.KiB(256)
-          vol.max_disk_size = DiskSize.MiB(8)
-          vol.desired_disk_size = DiskSize.MiB(1)
-          # Make sure that alignment does not result in a too big partition
-          vol.align = :keep_size
-          vol.bootable = true
-          vol.can_live_on_logical_volume = false
-          # TODO: We have been told that PReP must be one of the first 4
-          # partitions, ideally the first one. But we have not found any
-          # rational/evidence. Not implementing that for the time being
-          vol
-        end
+      def prep_volume
+        vol = PlannedVolume.new(nil)
+        # So far we are always using msdos partition ids
+        vol.partition_id = ::Storage::ID_PPC_PREP
+        vol.min_disk_size = DiskSize.KiB(256)
+        vol.max_disk_size = DiskSize.MiB(8)
+        vol.desired_disk_size = DiskSize.MiB(1)
+        # Make sure that alignment does not result in a too big partition
+        vol.align = :keep_size
+        vol.bootable = true
+        vol.can_live_on_logical_volume = false
+        # TODO: We have been told that PReP must be one of the first 4
+        # partitions, ideally the first one. But we have not found any
+        # rational/evidence. Not implementing that for the time being
+        vol
+      end
 
-        def arch
-          @arch ||= StorageManager.instance.arch
-        end
+      def arch
+        @arch ||= StorageManager.instance.arch
       end
     end
   end
