+++ src/lib/y2storage/devices_lists.rb
@@ -1,5 +1,3 @@
-#!/usr/bin/env ruby
-#
 # encoding: utf-8
 
 # Copyright (c) [2016] SUSE LLC
@@ -21,31 +19,7 @@
 # To contact SUSE LLC about this file by physical or electronic mail, you may
 # find current contact information at www.suse.com.
 
-require "storage"
-require "y2storage/disk_size"
-
-module Y2Storage
-  module Refinements
-    # Refinements for Partition adding some virtual attributes, mainly used
-    # to make the rspec tests more readable
-    module PartitionAttributes
-      refine ::Storage::Partition do
-        # First mounpoint
-        def mountpoint
-          filesystem.mountpoints.first
-        end
-
-        # Label of the filesystem
-        def label
-          filesystem.label
-        end
-
-        # UUID of the filesystem
-        def uuid
-          filesystem.uuid
-        end
-
-      end
-    end
-  end
-end
+require "y2storage/devices_lists/disks_list"
+require "y2storage/devices_lists/filesystems_list"
+require "y2storage/devices_lists/free_disk_spaces_list"
+require "y2storage/devices_lists/partitions_list"
