+++ src/lib/y2storage/proposal/devicegraph_generator.rb
@@ -22,126 +22,124 @@
 # find current contact information at www.suse.com.
 
 require "storage"
-require "storage/proposal/space_maker"
-require "storage/proposal/partition_creator"
-require "storage/refinements/devicegraph_lists"
+require "y2storage/proposal/space_maker"
+require "y2storage/proposal/partition_creator"
+require "y2storage/refinements/devicegraph_lists"
 
-module Yast
-  module Storage
-    class Proposal
-      # Class to create devicegraphs that can accommodate a given collection of
-      # volumes
-      class DevicegraphGenerator
-        include Yast::Logger
+module Y2Storage
+  class Proposal
+    # Class to create devicegraphs that can accommodate a given collection of
+    # volumes
+    class DevicegraphGenerator
+      include Yast::Logger
 
-        using Refinements::DevicegraphLists
+      using Refinements::DevicegraphLists
 
-        attr_accessor :settings
+      attr_accessor :settings
 
-        def initialize(settings)
-          @settings = settings
+      def initialize(settings)
+        @settings = settings
+      end
+
+      # Devicegraph including all the specified volumes
+      #
+      # @param volumes [PlannedVolumesList] volumes to accommodate
+      # @param initial_graph [::Storage::Devicegraph] initial devicegraph
+      #           (typically the representation of the current system)
+      # @param disk_analyzer [DiskAnalyzer] analysis of the initial_graph
+      #
+      # @return [::Storage::Devicegraph]
+      # @raise Proposal::Error if it was not possible to propose a devicegraph
+      def devicegraph(volumes, initial_graph, disk_analyzer)
+        # We are going to alter the volumes in several ways, so let's be a
+        # good citizen and do it in our own copy
+        volumes = volumes.deep_dup
+
+        space_maker = SpaceMaker.new(initial_graph, disk_analyzer, settings)
+        begin
+          space_result = provide_space(volumes, space_maker)
+        rescue NoDiskSpaceError
+          raise if volumes.target == :min
+          # Try again with the minimum size
+          volumes.target = :min
+          space_result = provide_space(volumes, space_maker)
         end
 
-        # Devicegraph including all the specified volumes
-        #
-        # @param volumes [PlannedVolumesList] volumes to accommodate
-        # @param initial_graph [::Storage::Devicegraph] initial devicegraph
-        #           (typically the representation of the current system)
-        # @param disk_analyzer [DiskAnalyzer] analysis of the initial_graph
-        #
-        # @return [::Storage::Devicegraph]
-        # @raise Proposal::Error if it was not possible to propose a devicegraph
-        def devicegraph(volumes, initial_graph, disk_analyzer)
-          # We are going to alter the volumes in several ways, so let's be a
-          # good citizen and do it in our own copy
-          volumes = volumes.deep_dup
+        refine_volumes!(volumes, space_result[:deleted_partitions])
+        graph = create_partitions(space_result[:space_distribution], space_result[:devicegraph])
+        reuse_partitions!(volumes, graph)
+        graph
+      end
 
-          space_maker = SpaceMaker.new(initial_graph, disk_analyzer, settings)
-          begin
-            space_result = provide_space(volumes, space_maker)
-          rescue NoDiskSpaceError
-            raise if volume.target == :min
-            # Try again with the minimum size
-            volumes.target = :min
-            space_result = provide_space(volumes, space_maker)
-          end
+    protected
 
-          refine_volumes!(volumes, space_result[:deleted_partitions])
-          graph = create_partitions(space_result[:space_distribution], space_result[:devicegraph])
-          reuse_partitions!(volumes, graph)
-          graph
+      # Provides free disk space in the proposal devicegraph to fit the
+      # volumes in.
+      #
+      # @raise Proposal::Error if the goal is not reached
+      #
+      # @param volumes [PlannedVolumesList] set of volumes to make space for
+      # @param space_maker [SpaceMaker]
+      #
+      # @return [::Storage::Devicegraph]
+      def provide_space(volumes, space_maker)
+        result = space_maker.provide_space(volumes)
+        log.info(
+          "Found #{volumes.target} space"
+        )
+        result
+      end
+
+      # Adds some extra information to the planned volumes inferred from
+      # the list of partitions deleted by the space maker.
+      #
+      # It enforces reuse of UUIDs and labels from the deleted swap
+      # partitions.
+      #
+      # It modifies the passed volumes.
+      #
+      # @param volumes [PlannedVolumesList] list of volumes to modify
+      # @param deleted_partitions [Array<::Storage::Partition>] partitions
+      #     deleted from the initial devicegraph
+      def refine_volumes!(volumes, deleted_partitions)
+        deleted_swaps = deleted_partitions.select do |part|
+          part.id == ::Storage::ID_SWAP
         end
+        new_swap_volumes = volumes.select { |vol| !vol.reuse && vol.mount_point == "swap" }
 
-      protected
+        new_swap_volumes.each_with_index do |swap_volume, idx|
+          deleted_swap = deleted_swaps[idx]
+          break unless deleted_swap
 
-        # Provides free disk space in the proposal devicegraph to fit the
-        # volumes in.
-        #
-        # @raise Proposal::Error if the goal is not reached
-        #
-        # @param volumes [PlannedVolumesList] set of volumes to make space for
-        # @param space_maker [SpaceMaker]
-        #
-        # @return [::Storage::Devicegraph]
-        def provide_space(volumes, space_maker)
-          result = space_maker.provide_space(volumes)
-          log.info(
-            "Found #{volumes.target} space"
-          )
-          result
+          swap_volume.uuid = deleted_swap.filesystem.uuid
+          swap_volume.label = deleted_swap.filesystem.label
         end
+      end
 
-        # Adds some extra information to the planned volumes inferred from
-        # the list of partitions deleted by the space maker.
-        #
-        # It enforces reuse of UUIDs and labels from the deleted swap
-        # partitions.
-        #
-        # It modifies the passed volumes.
-        #
-        # @param volumes [PlannedVolumesList] list of volumes to modify
-        # @param deleted_partitions [Array<::Storage::Partition>] partitions
-        #     deleted from the initial devicegraph
-        def refine_volumes!(volumes, deleted_partitions)
-          deleted_swaps = deleted_partitions.select do |part|
-            part.id == ::Storage::ID_SWAP
-          end
-          new_swap_volumes = volumes.select { |vol| !vol.reuse && vol.mount_point == "swap" }
+      # Creates partitions representing a set of volumes
+      #
+      # @param volumes [PlannedVolumesList] set of volumes to create
+      # @param initial_graph [::Storage::Devicegraph] initial devicegraph
+      #
+      # @return [::Storage::Devicegraph]
+      def create_partitions(distribution, initial_graph)
+        partition_creator = PartitionCreator.new(initial_graph)
+        partition_creator.create_partitions(distribution)
+      end
 
-          new_swap_volumes.each_with_index do |swap_volume, idx|
-            deleted_swap = deleted_swaps[idx]
-            break unless deleted_swap
-
-            swap_volume.uuid = deleted_swap.filesystem.uuid
-            swap_volume.label = deleted_swap.filesystem.label
-          end
-        end
-
-        # Creates partitions representing a set of volumes
-        #
-        # @param volumes [PlannedVolumesList] set of volumes to create
-        # @param initial_graph [::Storage::Devicegraph] initial devicegraph
-        #
-        # @return [::Storage::Devicegraph]
-        def create_partitions(distribution, initial_graph)
-          partition_creator = PartitionCreator.new(initial_graph)
-          partition_creator.create_partitions(distribution)
-        end
-
-        # Adjusts pre-existing (not created by us) partitions assigning its
-        # mount point and boot flag
-        #
-        # It works directly on the passed devicegraph
-        #
-        # @param volumes [PlannedVolumesList] set of volumes to create
-        # @param graph [::Storage::Devicegraph] devicegraph to modify
-        def reuse_partitions!(volumes, graph)
-          volumes.select { |v| v.reuse }.each do |vol|
-            partition = graph.partitions.with(name: vol.reuse).first
-            filesystem = partition.filesystem
-            filesystem.add_mountpoint(vol.mount_point) if vol.mount_point && !vol.mount_point.empty?
-            partition.boot = true if vol.bootable
-          end
+      # Adjusts pre-existing (not created by us) partitions assigning its
+      # mount point and boot flag
+      #
+      # It works directly on the passed devicegraph
+      #
+      # @param volumes [PlannedVolumesList] set of volumes to create
+      # @param graph [::Storage::Devicegraph] devicegraph to modify
+      def reuse_partitions!(volumes, graph)
+        volumes.select { |v| v.reuse }.each do |vol|
+          partition = graph.partitions.with(name: vol.reuse).first
+          filesystem = partition.filesystem
+          filesystem.add_mountpoint(vol.mount_point) if vol.mount_point && !vol.mount_point.empty?
+          partition.boot = true if vol.bootable
         end
       end
     end
