+++ src/lib/y2storage/proposal/space_maker.rb
@@ -23,360 +23,357 @@
 
 require "fileutils"
 require "storage"
-require "storage/planned_volume"
-require "storage/disk_size"
-require "storage/free_disk_space"
-require "storage/proposal/space_distribution"
-require "storage/refinements/devicegraph"
-require "storage/refinements/devicegraph_lists"
+require "y2storage/planned_volume"
+require "y2storage/disk_size"
+require "y2storage/free_disk_space"
+require "y2storage/proposal/space_distribution"
+require "y2storage/refinements"
 
-module Yast
-  module Storage
-    class Proposal
-      # Class to provide free space for creating new partitions - either by
-      # reusing existing unpartitioned space, by deleting existing partitions
-      # or by resizing an existing Windows partition.
-      class SpaceMaker
-        using Refinements::Devicegraph
-        using Refinements::DevicegraphLists
-        include Yast::Logger
+module Y2Storage
+  class Proposal
+    # Class to provide free space for creating new partitions - either by
+    # reusing existing unpartitioned space, by deleting existing partitions
+    # or by resizing an existing Windows partition.
+    class SpaceMaker
+      using Refinements::Devicegraph
+      using Refinements::DevicegraphLists
+      include Yast::Logger
 
-        attr_accessor :settings
+      attr_accessor :settings
 
-        # Initialize.
-        #
-        # @param original_graph [::Storage::Devicegraph] initial devicegraph
-        # @param disk_analyzer [DiskAnalyzer] information about original_graph
-        # @param settings [Proposal::Settings] proposal settings
-        def initialize(original_graph, disk_analyzer, settings)
-          @original_graph = original_graph
-          @disk_analyzer = disk_analyzer
-          @settings = settings
+      # Initialize.
+      #
+      # @param original_graph [::Storage::Devicegraph] initial devicegraph
+      # @param disk_analyzer [DiskAnalyzer] information about original_graph
+      # @param settings [ProposalSettings] proposal settings
+      def initialize(original_graph, disk_analyzer, settings)
+        @original_graph = original_graph
+        @disk_analyzer = disk_analyzer
+        @settings = settings
+      end
+
+      # Performs all the operations needed to free enough space to accomodate
+      # a set of volumes
+      #
+      # @raise Proposal::Error if is not possible to accomodate the volumes
+      #
+      # @param volumes [PlannedVolumesList] volumes to make space for
+      # @return [Hash] a hash with three elements:
+      #   devicegraph: [::Storage::Devicegraph] resulting devicegraph
+      #   deleted_partitions: [Array<::Storage::Partition>] partitions that
+      #     were in the original devicegraph but are not in the resulting one
+      #   space_distribution: [SpaceDistribution] proposed distribution of
+      #     volumes
+      #
+      def provide_space(volumes)
+        @new_graph = original_graph.copy
+        @deleted_names = []
+
+        # Partitions that should not be deleted
+        keep = []
+        # Let's filter out volumes with some value in #reuse
+        volumes = volumes.dup
+        volumes.select(&:reuse).each do |vol|
+          log.info "No need to find a fit for this volume, it will reuse #{vol.reuse}: #{vol}"
+          keep << vol.reuse
+          volumes.delete(vol)
         end
 
-        # Performs all the operations needed to free enough space to accomodate
-        # a set of volumes
-        #
-        # @raise Proposal::Error if is not possible to accomodate the volumes
-        #
-        # @param volumes [PlannedVolumesList] volumes to make space for
-        # @return [Hash] a hash with three elements:
-        #   devicegraph: [::Storage::Devicegraph] resulting devicegraph
-        #   deleted_partitions: [Array<::Storage::Partition>] partitions that
-        #     were in the original devicegraph but are not in the resulting one
-        #   space_distribution: [SpaceDistribution] proposed distribution of
-        #     volumes
-        #
-        def provide_space(volumes)
-          @new_graph = original_graph.copy
-          @deleted_names = []
-
-          # Partitions that should not be deleted
-          keep = []
-          # Let's filter out volumes with some value in #reuse
-          volumes = volumes.dup
-          volumes.select(&:reuse).each do |vol|
-            log.info "No need to find a fit for this volume, it will reuse #{vol.reuse}: #{vol}"
-            keep << vol.reuse
-            volumes.delete(vol)
-          end
-
-          # To make sure we are not freeing space in a useless place, let's
-          # first restrict the operations to disks with particular requirements
-          volumes_by_disk(volumes).each do |disk, vols|
-            resize_and_delete!(vols, keep, disk: disk)
-          end
-          # Doing something similar for #max_start_offset is more difficult and
-          # doesn't pay off (#max_start_offset is used just in one case)
-
-          # Now with the full set of volumes and all the candidate disks
-          resize_and_delete!(volumes, keep)
-
-          {
-            devicegraph:        @new_graph,
-            deleted_partitions: deleted_partitions,
-            space_distribution: @distribution
-          }
+        # To make sure we are not freeing space in a useless place, let's
+        # first restrict the operations to disks with particular requirements
+        volumes_by_disk(volumes).each do |disk, vols|
+          resize_and_delete!(vols, keep, disk: disk)
         end
+        # Doing something similar for #max_start_offset is more difficult and
+        # doesn't pay off (#max_start_offset is used just in one case)
 
-      protected
+        # Now with the full set of volumes and all the candidate disks
+        resize_and_delete!(volumes, keep)
 
-        attr_reader :original_graph, :new_graph, :disk_analyzer
+        {
+          devicegraph:        @new_graph,
+          deleted_partitions: deleted_partitions,
+          space_distribution: @distribution
+        }
+      end
 
-        # Partitions from the original devicegraph that are not present in the
-        # result of the last call to #provide_space
-        #
-        # @return [Array<::Storage::Partition>]
-        def deleted_partitions
-          original_graph.partitions.with(name: @deleted_names).to_a
-        end
+    protected
 
-        # @return [Hash{String => PlannedVolumesList}]
-        def volumes_by_disk(volumes)
-          volumes.each_with_object({}) do |volume, hash|
-            if volume.disk
-              hash[volume.disk] ||= PlannedVolumesList.new([], target: volumes.target)
-              hash[volume.disk] << volume
-            end
+      attr_reader :original_graph, :new_graph, :disk_analyzer
+
+      # Partitions from the original devicegraph that are not present in the
+      # result of the last call to #provide_space
+      #
+      # @return [Array<::Storage::Partition>]
+      def deleted_partitions
+        original_graph.partitions.with(name: @deleted_names).to_a
+      end
+
+      # @return [Hash{String => PlannedVolumesList}]
+      def volumes_by_disk(volumes)
+        volumes.each_with_object({}) do |volume, hash|
+          if volume.disk
+            hash[volume.disk] ||= PlannedVolumesList.new([], target: volumes.target)
+            hash[volume.disk] << volume
           end
         end
+      end
 
-        # Checks whether the goal has already being reached
-        #
-        # If it returns true, it stores in @distribution the SpaceDistribution
-        # that made it possible.
-        #
-        # @return [Boolean]
-        def success?(volumes)
-          @distribution ||= SpaceDistribution.best_for(volumes, free_spaces(new_graph).to_a, new_graph)
-          !!@distribution
-        rescue Error => e
-          log.info "Exception while trying to distribute volumes: #{e}"
-          @distribution = nil
-          false
+      # Checks whether the goal has already being reached
+      #
+      # If it returns true, it stores in @distribution the SpaceDistribution
+      # that made it possible.
+      #
+      # @return [Boolean]
+      def success?(volumes)
+        @distribution ||= SpaceDistribution.best_for(volumes, free_spaces(new_graph).to_a, new_graph)
+        !!@distribution
+      rescue Error => e
+        log.info "Exception while trying to distribute volumes: #{e}"
+        @distribution = nil
+        false
+      end
+
+      # Perform all the needed operations to make space for the volumes
+      #
+      # @param volumes [PlannedVolumesList] volumes to make space for
+      # @param keep [Array<String>] device names of partitions that should not
+      #     be deleted
+      # @param disk [String] optional disk name to restrict operations to
+      def resize_and_delete!(volumes, keep, disk: nil)
+        log.info "Resize and delete. disk: #{disk}"
+
+        @distribution = nil
+
+        # Initially, resize only if there are no Linux partitions in the disk
+        resize_windows!(volumes, disk, force: false)
+        delete_partitions!(volumes, :linux, keep, disk)
+        delete_partitions!(volumes, :other, keep, disk)
+        # If everything else failed, try resizing Windows before deleting it
+        resize_windows!(volumes, disk, force: true)
+        delete_partitions!(volumes, :windows, keep, disk)
+
+        raise NoDiskSpaceError unless success?(volumes)
+      end
+
+      # Additional space that needs to be freed in order to reach the goal
+      #
+      # @return [DiskSize]
+      def missing_required_size(volumes, disk)
+        SpaceDistribution.missing_disk_size(volumes, free_spaces(new_graph, disk).to_a)
+      end
+
+      # List of free spaces in the given devicegraph
+      #
+      # @param graph [::Storage::Devicegraph]
+      # @param disk [String] optional disk name to restrict result to
+      # @return [FreeDiskSpacesList]
+      def free_spaces(graph, disk = nil)
+        disks_for(graph, disk).free_disk_spaces
+      end
+
+      # List of candidate disks in the given devicegraph
+      #
+      # @param devicegraph [::Storage::Devicegraph]
+      # @param disk [String] optional disk name to restrict result to
+      # @return [DisksList]
+      def disks_for(devicegraph, disk = nil)
+        filter = disk || candidate_disk_names
+        devicegraph.disks.with(name: filter)
+      end
+
+      # @return [Array<String>]
+      def candidate_disk_names
+        settings.candidate_devices
+      end
+
+      # Try to resize the existing windows partitions
+      #
+      # @param volumes [PlannedVolumesList] list of volumes to allocate, used
+      #   to know how much space is still missing
+      # @param disk [String] optional disk name to restrict operations to
+      # @param force [Boolean] whether to resize Windows even if there are
+      #   Linux partitions in the same disk
+      def resize_windows!(volumes, disk, force: false)
+        return if success?(volumes)
+        part_names = windows_part_names(disk)
+        return if part_names.empty?
+
+        log.info("Resizing Windows partitions (force: #{force}")
+        parts_by_disk = partitions_by_disk(part_names)
+        remove_linux_disks!(parts_by_disk) unless force
+
+        success = sorted_resizables(parts_by_disk.values.flatten).any? do |res|
+          shrink_size = [
+            res[:recoverable_size],
+            missing_required_size(volumes, disk)
+          ].min
+          shrink_partition(res[:partition], shrink_size)
+
+          success?(volumes)
         end
 
-        # Perform all the needed operations to make space for the volumes
-        #
-        # @param volumes [PlannedVolumesList] volumes to make space for
-        # @param keep [Array<String>] device names of partitions that should not
-        #     be deleted
-        # @param disk [String] optional disk name to restrict operations to
-        def resize_and_delete!(volumes, keep, disk: nil)
-          log.info "Resize and delete. disk: #{disk}"
+        log.info "Didn't manage to free enough space by resizing Windows" unless success
+      end
 
-          @distribution = nil
-
-          # Initially, resize only if there are no Linux partitions in the disk
-          resize_windows!(volumes, disk, force: false)
-          delete_partitions!(volumes, :linux, keep, disk)
-          delete_partitions!(volumes, :other, keep, disk)
-          # If everything else failed, try resizing Windows before deleting it
-          resize_windows!(volumes, disk, force: true)
-          delete_partitions!(volumes, :windows, keep, disk)
-
-          raise NoDiskSpaceError unless success?(volumes)
+      # @return [Hash{String => ::Storage::Partition}]
+      def partitions_by_disk(part_names)
+        partitions = new_graph.partitions.with(name: part_names)
+        partitions.each_with_object({}) do |partition, hash|
+          disk_name = partition.partitionable.name
+          hash[disk_name] ||= []
+          hash[disk_name] << partition
         end
+      end
 
-        # Additional space that needs to be freed in order to reach the goal
-        #
-        # @return [DiskSize]
-        def missing_required_size(volumes, disk)
-          SpaceDistribution.missing_disk_size(volumes, free_spaces(new_graph, disk).to_a)
-        end
-
-        # List of free spaces in the given devicegraph
-        #
-        # @param graph [::Storage::Devicegraph]
-        # @param disk [String] optional disk name to restrict result to
-        # @return [FreeDiskSpacesList]
-        def free_spaces(graph, disk = nil)
-          disks_for(graph, disk).free_disk_spaces
-        end
-
-        # List of candidate disks in the given devicegraph
-        #
-        # @param devicegraph [::Storage::Devicegraph]
-        # @param disk [String] optional disk name to restrict result to
-        # @return [DisksList]
-        def disks_for(devicegraph, disk = nil)
-          filter = disk || candidate_disk_names
-          devicegraph.disks.with(name: filter)
-        end
-
-        # @return [Array<String>]
-        def candidate_disk_names
-          settings.candidate_devices
-        end
-
-        # Try to resize the existing windows partitions
-        #
-        # @param volumes [PlannedVolumesList] list of volumes to allocate, used
-        #   to know how much space is still missing
-        # @param disk [String] optional disk name to restrict operations to
-        # @param force [Boolean] whether to resize Windows even if there are
-        #   Linux partitions in the same disk
-        def resize_windows!(volumes, disk, force: false)
-          return if success?(volumes)
-          part_names = windows_part_names(disk)
-          return if part_names.empty?
-
-          log.info("Resizing Windows partitions (force: #{force}")
-          parts_by_disk = partitions_by_disk(part_names)
-          remove_linux_disks!(parts_by_disk) unless force
-
-          success = sorted_resizables(parts_by_disk.values.flatten).any? do |res|
-            shrink_size = [
-              res[:recoverable_size],
-              missing_required_size(volumes, disk)
-            ].min
-            shrink_partition(res[:partition], shrink_size)
-
-            success?(volumes)
-          end
-
-          log.info "Didn't manage to free enough space by resizing Windows" unless success
-        end
-
-        # @return [Hash{String => ::Storage::Partition}]
-        def partitions_by_disk(part_names)
-          partitions = new_graph.partitions.with(name: part_names)
-          partitions.each_with_object({}) do |partition, hash|
-            disk_name = partition.partitionable.name
-            hash[disk_name] ||= []
-            hash[disk_name] << partition
-          end
-        end
-
-        def remove_linux_disks!(partitions_by_disk)
-          partitions_by_disk.each do |disk, _p|
-            if linux_in_disk?(disk)
-              log.info "Linux partitions in #{disk}, Windows will not be resized"
-              partitions_by_disk.delete(disk)
-            end
-          end
-        end
-
-        def linux_in_disk?(disk_name)
-          linux_part_names(disk_name).any?
-        end
-
-        # List of partitions that can be resized, including the size of the
-        # space that can be reclaimed for each partition.
-        #
-        # The list is sorted so the partitions with more recoverable space are
-        # listed first.
-        #
-        # @param partitions [Array<::Storage::Partition>] list of partitions
-        # @return [Array<Hash>] each element contains
-        #     :partition (::Storage::Partition) and :recoverable_size (DiskSize)
-        def sorted_resizables(partitions)
-          resizables = partitions.map do |part|
-            { partition: part, recoverable_size: recoverable_size(part) }
-          end
-
-          resizables.delete_if { |res| res[:recoverable_size].zero? }
-          resizables.sort_by { |res| res[:recoverable_size] }.reverse
-        end
-
-        # Size of the space that can be reclaimed in a partition
-        #
-        # @param partition [::Storage::Partition]
-        # @return [DiskSize]
-        def recoverable_size(partition)
-          info = partition.filesystem.detect_resize_info
-          return DiskSize.zero unless info.resize_ok
-          DiskSize.B(partition.size - info.min_size)
-        end
-
-        # Reduces the size of a partition
-        #
-        # @param partition [::Storage::Partition]
-        # @param shrink_size [DiskSize] size of the space to substract
-        def shrink_partition(partition, shrink_size)
-          log.info "Shrinking #{partition.name} by #{shrink_size}"
-          partition.size = partition.size - shrink_size.to_i
-        end
-
-        # Use force to create space: delete partitions if a given type while
-        # there is no suitable space distribution
-        #
-        # @see #deletion_candidate_partitions for supported types
-        #
-        # @param volumes [PlannedVolumesList]
-        # @param type [Symbol] type of partition to delete
-        # @param keep [Array<String>] device names of partitions that should not
-        #       be deleted
-        # @param disk [String] optional disk name to restrict operations to
-        def delete_partitions!(volumes, type, keep, disk)
-          return if success?(volumes)
-
-          log.info("Deleting partitions to make space")
-          deletion_candidate_partitions(type, disk).each do |part_name|
-            if keep.include?(part_name)
-              log.info "Skipped deletion of #{part_name}"
-              next
-            end
-            part = find_partition(part_name)
-            next unless part
-            delete_partition(part)
-            break if success?(volumes)
+      def remove_linux_disks!(partitions_by_disk)
+        partitions_by_disk.each do |disk, _p|
+          if linux_in_disk?(disk)
+            log.info "Linux partitions in #{disk}, Windows will not be resized"
+            partitions_by_disk.delete(disk)
           end
         end
+      end
 
-        def find_partition(name)
-          ::Storage::Partition.find_by_name(new_graph, name)
-        rescue
-          nil
+      def linux_in_disk?(disk_name)
+        linux_part_names(disk_name).any?
+      end
+
+      # List of partitions that can be resized, including the size of the
+      # space that can be reclaimed for each partition.
+      #
+      # The list is sorted so the partitions with more recoverable space are
+      # listed first.
+      #
+      # @param partitions [Array<::Storage::Partition>] list of partitions
+      # @return [Array<Hash>] each element contains
+      #     :partition (::Storage::Partition) and :recoverable_size (DiskSize)
+      def sorted_resizables(partitions)
+        resizables = partitions.map do |part|
+          { partition: part, recoverable_size: recoverable_size(part) }
         end
 
-        # Deletes a given partition from its corresponding partition table
-        def delete_partition(partition)
-          log.info("Deleting partition #{partition.name} in device graph")
-          @deleted_names << partition.name
-          partition.partition_table.delete_partition(partition.name)
-        end
+        resizables.delete_if { |res| res[:recoverable_size].zero? }
+        resizables.sort_by { |res| res[:recoverable_size] }.reverse
+      end
 
-        # Partitions of a given type to be deleted. The type can be:
-        #
-        #  * :windows Partitions with a Windows installation on it
-        #  * :linux Partitions that are part of a Linux installation
-        #  * :other Any other partition
-        #
-        # @param type [Symbol]
-        # @param disk [String] optional disk name to restrict operations to
-        # @return [Array<String>] partition names sorted by disk and by position
-        #     inside the disk (partitions at the end are presented first)
-        def deletion_candidate_partitions(type, disk = nil)
-          names = []
-          disks_for(original_graph, disk).each do |dsk|
-            partitions = original_graph.disks.with(name: dsk.name).partitions.to_a
-            filter_partitions_by_type!(partitions, type, dsk.name)
-            partitions = partitions.sort_by { |part| part.region.start }.reverse
-            names += partitions.map(&:name)
+      # Size of the space that can be reclaimed in a partition
+      #
+      # @param partition [::Storage::Partition]
+      # @return [DiskSize]
+      def recoverable_size(partition)
+        info = partition.filesystem.detect_resize_info
+        return DiskSize.zero unless info.resize_ok
+        DiskSize.B(partition.size - info.min_size)
+      end
+
+      # Reduces the size of a partition
+      #
+      # @param partition [::Storage::Partition]
+      # @param shrink_size [DiskSize] size of the space to substract
+      def shrink_partition(partition, shrink_size)
+        log.info "Shrinking #{partition.name} by #{shrink_size}"
+        partition.size = partition.size - shrink_size.to_i
+      end
+
+      # Use force to create space: delete partitions if a given type while
+      # there is no suitable space distribution
+      #
+      # @see #deletion_candidate_partitions for supported types
+      #
+      # @param volumes [PlannedVolumesList]
+      # @param type [Symbol] type of partition to delete
+      # @param keep [Array<String>] device names of partitions that should not
+      #       be deleted
+      # @param disk [String] optional disk name to restrict operations to
+      def delete_partitions!(volumes, type, keep, disk)
+        return if success?(volumes)
+
+        log.info("Deleting partitions to make space")
+        deletion_candidate_partitions(type, disk).each do |part_name|
+          if keep.include?(part_name)
+            log.info "Skipped deletion of #{part_name}"
+            next
           end
+          part = find_partition(part_name)
+          next unless part
+          delete_partition(part)
+          break if success?(volumes)
+        end
+      end
 
-          log.info "Deletion candidates (#{type}): #{names}"
-          names
+      def find_partition(name)
+        ::Storage::Partition.find_by_name(new_graph, name)
+      rescue
+        nil
+      end
+
+      # Deletes a given partition from its corresponding partition table
+      def delete_partition(partition)
+        log.info("Deleting partition #{partition.name} in device graph")
+        @deleted_names << partition.name
+        partition.partition_table.delete_partition(partition.name)
+      end
+
+      # Partitions of a given type to be deleted. The type can be:
+      #
+      #  * :windows Partitions with a Windows installation on it
+      #  * :linux Partitions that are part of a Linux installation
+      #  * :other Any other partition
+      #
+      # @param type [Symbol]
+      # @param disk [String] optional disk name to restrict operations to
+      # @return [Array<String>] partition names sorted by disk and by position
+      #     inside the disk (partitions at the end are presented first)
+      def deletion_candidate_partitions(type, disk = nil)
+        names = []
+        disks_for(original_graph, disk).each do |dsk|
+          partitions = original_graph.disks.with(name: dsk.name).partitions.to_a
+          filter_partitions_by_type!(partitions, type, dsk.name)
+          partitions = partitions.sort_by { |part| part.region.start }.reverse
+          names += partitions.map(&:name)
         end
 
-        def filter_partitions_by_type!(partitions, type, disk)
-          case type
-          when :windows
-            partitions.select! { |part| windows_part_names(disk).include?(part.name) }
-          when :linux
-            partitions.select! { |part| linux_part_names(disk).include?(part.name) }
-          when :other
-            partitions.select! do |part|
-              !linux_part_names(disk).include?(part.name) &&
-                !windows_part_names(disk).include?(part.name)
-            end
+        log.info "Deletion candidates (#{type}): #{names}"
+        names
+      end
+
+      def filter_partitions_by_type!(partitions, type, disk)
+        case type
+        when :windows
+          partitions.select! { |part| windows_part_names(disk).include?(part.name) }
+        when :linux
+          partitions.select! { |part| linux_part_names(disk).include?(part.name) }
+        when :other
+          partitions.select! do |part|
+            !linux_part_names(disk).include?(part.name) &&
+              !windows_part_names(disk).include?(part.name)
           end
         end
+      end
 
-        # Device names of windows partitions detected by disk_analyzer
-        #
-        # @return [array<string>]
-        def windows_part_names(disk = nil)
-          parts = if disk
-            disk_analyzer.windows_partitions[disk] || []
-          else
-            disk_analyzer.windows_partitions.values.flatten
-          end
-          parts.map(&:name)
+      # Device names of windows partitions detected by disk_analyzer
+      #
+      # @return [array<string>]
+      def windows_part_names(disk = nil)
+        parts = if disk
+          disk_analyzer.windows_partitions[disk] || []
+        else
+          disk_analyzer.windows_partitions.values.flatten
         end
+        parts.map(&:name)
+      end
 
-        # Device names of linux partitions detected by disk_analyzer
-        #
-        # @return [array<string>]
-        def linux_part_names(disk = nil)
-          parts = if disk
-            disk_analyzer.linux_partitions[disk] || []
-          else
-            disk_analyzer.linux_partitions.values.flatten
-          end
-          parts.map(&:name)
+      # Device names of linux partitions detected by disk_analyzer
+      #
+      # @return [array<string>]
+      def linux_part_names(disk = nil)
+        parts = if disk
+          disk_analyzer.linux_partitions[disk] || []
+        else
+          disk_analyzer.linux_partitions.values.flatten
         end
+        parts.map(&:name)
       end
     end
   end
