+++ src/lib/y2storage/refinements/size_casts.rb
@@ -22,30 +22,28 @@
 # find current contact information at www.suse.com.
 
 require "storage"
-require "storage/disk_size"
+require "y2storage/disk_size"
 
-module Yast
-  module Storage
-    module Refinements
-      # Refinements to make DiskSize objects more readable in the rspec tests
-      #
-      # It adds methods to perform a direct cast from numerical classes into
-      # DiskSize objects
-      # @example
-      #   using Yast::Storage::Refinements::SizeCasts
-      #
-      #   20.GiB == Yast::Storage::DiskSize.GiB(20)
-      #   12.5.MiB == Yast::Storage::DiskSize.MiB(12.5)
-      module SizeCasts
-        REFINED_CLASSES = [::Fixnum, ::Float]
-        ADDED_METHODS = [:KiB, :MiB, :GiB, :TiB, :PiB, :EiB, :ZiB, :YiB]
+module Y2Storage
+  module Refinements
+    # Refinements to make DiskSize objects more readable in the rspec tests
+    #
+    # It adds methods to perform a direct cast from numerical classes into
+    # DiskSize objects
+    # @example
+    #   using Y2Storage::Refinements::SizeCasts
+    #
+    #   20.GiB == Y2Storage::DiskSize.GiB(20)
+    #   12.5.MiB == Y2Storage::DiskSize.MiB(12.5)
+    module SizeCasts
+      REFINED_CLASSES = [::Fixnum, ::Float]
+      ADDED_METHODS = [:KiB, :MiB, :GiB, :TiB, :PiB, :EiB, :ZiB, :YiB]
 
-        REFINED_CLASSES.each do |klass|
-          refine klass do
-            ADDED_METHODS.each do |method|
-              define_method(method) do
-                DiskSize.send(method, self)
-              end
+      REFINED_CLASSES.each do |klass|
+        refine klass do
+          ADDED_METHODS.each do |method|
+            define_method(method) do
+              DiskSize.send(method, self)
             end
           end
         end
